src/python/requirements.txt

flask==2.0.1
flask-cors==3.0.10
opencv-python==4.5.3.56
numpy==1.21.2
pillow==8.3.1
